requirements.txt
streamlit>=1.36.0
openai>=1.35.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
